requirements.txt
ollama>=0.6.1
fastmcp>=2.13.1
PyYAML>=6.0.3
python-dotenv>=1.0.0
